# backend/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: LingCast 数字人直播平台 API（api-admin / api-user / api-scheduler 共用一份 OpenAPI
    文档；管理端写操作与 /users、/admin/* 需登录会话，观众端与 Worker Webhook 公开）。
  title: LingCast API
  version: 1.0.0
paths:
...
  /chat/history:
    get:
      parameters:
      - description: Filter by avatar
        in: query
        name: avatarId
        type: integer
      - description: Filter by user
        in: query
        name: userId
        type: integer
      - description: Page (1-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Persisted room chat history (paginated)
      tags:
      - chat
  /chat/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: login form
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login (merges guest messages into the account)
      tags:
      - chat
  /chat/logs:
    get:
      parameters:
      - description: Filter by avatar
        in: query
        name: avatarId
        type: integer
      - description: Filter by user
        in: query
        name: userId
        type: integer
      - description: Filter by date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      - description: Keyword in message content
        in: query
        name: q
        type: string
      - description: Page (1-based)
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Admin chat logs (avatarId/userId/date/q + pagination)
      tags:
      - admin
  /chat/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: registration form
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Register (upgrades the guest row, keeps history)
      tags:
      - chat
...
swagger: "2.0"